BT076 - feat: get all hashtags
Edit method has been moved to StandardUserController
